Pull domain checking into utils
Previously, buyer domain checking was an RPC to the API.  Now it's just
a function in the utils.  The core functionality is just copied as-is
from the API code.

The list of valid domains couldn't be updated without a restart of the API
anyway.  It's theoretically possible to add live updating support, but
we probably won't because we won't need it.  We can do zero-downtime
updates to the frontend code anyway, so this approach gives us simpler
code and less stuff to mock in test cases.

I also took the opportunity to give the email validation routines more
descriptive and consistent names.