Handle BASE_TEMPLATE_DATA in the utils
Currently all the frontends are handling this separately in their
configs.  But this stops ASSET_PATH being handled correctly if it's
set in an environment variable.

All the frontends are copying the same code for this anyway, and if one
frontend needs a new variable, there's probably no harm letting other frontends
have it, too.  In the worst case, extra variables can be injected per
frontend if really needed.